scripts/spark_streaming/requirements.txt
pyspark==4.1.1
clickhouse-connect
python-dotenv